requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
beautifulsoup4
requests
duckduckgo-search
googlesearch-python
crawl4ai
playwright
